Changed PhoneNumbers to be a list of longs instead of ints

+++ CassandraFinal/Contact.cs
@@ -10,14 +10,14 @@ namespace CassandraFinal {
 
         public Guid ID { get; set; }
         public string Name { get; set; }
-        public List<int> PhoneNumbers { get; set; }
+        public List<long> PhoneNumbers { get; set; }
         public List<string> Emails { get; set; }
         public string Group { get; set; }
 
         // Used by Cassandra driver to populate properties
         public Contact() { }
 
-        public Contact(string name, List<int> phoneNumbers, List<string> emails, string group) {
+        public Contact(string name, List<long> phoneNumbers, List<string> emails, string group) {
             ID = Guid.NewGuid();
             Name = name;
             PhoneNumbers = phoneNumbers;
@@ -27,7 +27,7 @@ namespace CassandraFinal {
 
         public override string ToString() {
             string contact = $"ID: {ID}\nName: {Name}Group: {Group}\nPhone numbers:";
-            foreach (int number in PhoneNumbers) {
+            foreach (long number in PhoneNumbers) {
                 contact += $"\n\t{number}";
             }
             contact += "\nEmails:";
